Move newsletter signup route handler
